fix(patches): address Copilot review comments
- Fix _no_lora_dropout to return a proper bool (was returning None for
  non-None projections, disabling QKV/MLP fusion unconditionally)
- Cast A_i/B_i to x.dtype in multi-adapter loop to avoid dtype mismatch
- Add is_transformers_patched() to opaque.patches and re-export from
  opaque.patches.transformers.components (fixes ImportError in examples)
- Fix _patch_forward instance-level idempotency check to use __func__
  when inspecting bound method marker
- Replace print() with logger.warning() in _router.py
- Fix apply_kernels_for() to actually use the family argument instead
  of re-running auto-detection
- Move apply_model_patches() before get_peft_model() in both example
  scripts so detect_family() inspects the base Transformers config
- Update mkdocs.yml paths: opaque-performance/src -> opaque-patches/src
- Update NOTICE file paths to opaque-patches/src/opaque/patches/kernels/
- Fix docs/limitations.md: opaque.transformers.patches -> opaque.patches

+++ packages/opaque-patches/src/opaque/patches/peft/components/linear.py
@@ -65,8 +65,8 @@ def _make_lora_linear_forward(original):
             if adapter in self.lora_A:
                 dropout_i = self.lora_dropout[adapter]
                 x_i = dropout_i(x)
-                A_i = self.lora_A[adapter].weight.T
-                B_i = self.lora_B[adapter].weight.T
+                A_i = self.lora_A[adapter].weight.T.to(x.dtype)
+                B_i = self.lora_B[adapter].weight.T.to(x.dtype)
                 scaling_i = self.scaling[adapter]
                 lora_out = (x_i @ A_i) @ B_i * scaling_i
                 result = result + lora_out
